QAPI: new QMP command option "need-bql"
Introducing this new parameter for QMP commands in general to mark out
whether the command needs BQL during execution.  Normally QMP command
executions are done with the protection of BQL in QEMU.  However the
truth is that not all the QMP commands require the BQL.

This new parameter provides a way to allow QMP commands to run in
parallel when possible, without taking BQL before hand.

Since the default value of "need-bql" is true, so now all QMP commands
are still protected by BQL.

Signed-off-by: Peter Xu <peterx@redhat.com>