monitor: allow monitor to create thread to poll
Firstly, introduce Monitor.use_thread, and set it for monitors that are
using non-mux typed backend chardev.  We only do this for monitors, so
mux-typed chardevs are not suitable (when it connects to, e.g., serials
and the monitor together).

When use_thread is set, we create standalone thread to poll the monitor
events, isolated from the main loop thread.  Here we still need to take
the BQL before dispatching the tasks since some of the monitor commands
are not allowed to execute without the protection of BQL.  Then this
gives us the chance to avoid taking the BQL for some monitor commands in
the future.

* Why this change?

We need these per-monitor threads to make sure we can have at least one
monitor that will never block by BQL holders (that can receive further
monitor commands).

* So when will monitors get stuck?  And, how do they get stuck?

After we have postcopy and remote page faults, it's simple to let the
monitor stuck (which will also get main loop stuck):

(1) Monitor deadlock on BQL

As we may know, when postcopy is running on destination VM, the vcpu
threads can stop running merely any time as long as it tries to access
an uncopied guest page.  Meanwhile, when that happens, it is possible
that the vcpu thread is holding the BQL.  If the page fault is not
handled quickly, you'll find that monitors stop working, which is trying
to take the BQL.

If the page fault cannot be handled correctly (one case is a paused
postcopy, when network is temporarily down), monitors will hang
forever.  Without current patch, that means the main loop hanged.  We'll
never find a way to talk to VM again.

(2) Monitor tries to run codes page-faulted vcpus

The HMP command "info cpus" is one of the good example - it tries to
kick all the vcpus and sync status from them.  However, if there is any
vcpu that is stuck at an unhandled page fault, it can never achieve the
sync, then the HMP hangs.  Again, it hangs the main loop thread as well.

After either (1) or (2), we can see the deadlock problem:

- On one hand, if monitor hangs, we cannot do the postcopy recovery,
    because postcopy recovery needs user to specify new listening port on
    destination monitor.

- On the other hand, if we cannot recover the paused postcopy, then page
    faults cannot be serviced, and the monitors will possibly hang
    forever then.

* How this patch helps?

- Firstly, we'll have our own thread for each dedicated monitor (or say,
    the backend chardev is only used for monitor), so even main loop
    thread hangs (it is always possible), this monitor thread may still
    survive.

- Not all monitor commands need the BQL.  We can selectively take the
    BQL (depends on which command we are running) to avoid waiting on a
    page-faulted vcpu thread that has taken the BQL (this will be done in
    following up patches).

Signed-off-by: Peter Xu <peterx@redhat.com>