qmp: add subcmd "sleep-100ms" for x-oob-test
It can be used even without OOB enabled, to halt the main thread for a
while (the most important thing is, I never can have a command that can
halt the main thread statically on every single architecture...).  With
this, we can add some more strict tests for QMP.

Why 20ms? That's a vague number from me purely for testing, long enough
for schedulers to switch things in and out (assuming the old ticks are
100HZ, and that's a doubled value), and merely undetectable for human
beings who run the tests.  It's a stupid number and command, but I hope
that works efficiently to detect possible problems.

Signed-off-by: Peter Xu <peterx@redhat.com>