tests: fix iotests 109, 183 by ignore RESUME
When QEMU starts, it's possible that we'll get a RESUME event, depending
on how fast we connect the client to the QMP server.  If we are fast
enough, we'll be lucky to have that.

But it's unlucky that it can be undeterministic.  Let's be deterministic
by ignoring all RESUME event.

Test 183 will definitely receive a RESUME message since there is a
migration.  Remove that too since the test should not depend on that.

Signed-off-by: Peter Xu <peterx@redhat.com>