tests: fix iotest 205 on export remove failure
We'll conditionally get "export 'exp' still in use" and fail this test.
Fix that temporarily by waiting for a while before removing.  The ideal
solution would be having some way to query the status of NBD server
before removing exports.  However it seems that we still don't have
that.

My thanks to Kevin Wolf who helped during investinating the problem.

Signed-off-by: Peter Xu <peterx@redhat.com>