chardev: introduce chardev contexts
Introduce contexts in chardev layer.  Normally chardevs are running with
the main context except the monitors.  Let's create the first standalone
context for the monitors.  Note that this does not mean the chardev
backend will depend on the monitor code, but it's just a naming that let
people know this context should only be used by monitors.

This patch itself should not have any functional change (we'll create a
new iothread that managed by chardev module, but no one is currently
using it), however after this patch we should be able to specify the
context for a chardev by using a qemu_opt_set_number() onto the QemuOps
before the chardev backend is created.

Signed-off-by: Peter Xu <peterx@redhat.com>