migration: Postcopy preemption on partial huge page sending
This patch enables postcopy-preempt feature during partial page sending.

Before this patch, postcopy requests need to wait if precopy is sending one
huge page until the whole huge page is sent.  After this patch, postcopy can
preempt an huge page sending process of precopy so we can quickly switch to
sending the postcopy page and then we'll go back to continue sending the
precopy page after we serviced all the urgent pages.

Firstly, this feature will not be enabled if using 4K pages across the whole
guest memory.  It's only useful for preempt sending of a partial huge page or
it'll be pure overhead.

Dest QEMU should have already be okay on having more than one channel to
receive huge pages during postcopy.  This patch is focusing on the src QEMU to
decide when to preempt the precopy stream, maintaining the preempted status and
restore the precopy stream when we finished all postcopy requests.

After this patch, we'll have two "channels" for postcopy: (1) PRECOPY
channel (which is the default channel), and (2) POSTCOPY channel.  When we
found a postcopy request, we'll interrupt the PRECOPY channel sending process
and quickly switch to the POSTCOPY channel.  After we serviced all the high
priority postcopy pages, we'll switch back to PRECOPY channel so that we'll
continue to send the interrupted huge page again.

NOTE: this patch has a side effect on that after sending postcopy pages,
previously we'll assume guest will access follow up pages so we'll keep sending
from there.  Now it's not; instead we'll go back and continue sending the
interrupted precopy huge page.  Whether that's a problem is debatable, because
"assuming the guest will continue to access the next page" doesn't really suite
when huge pages are used, especially if the huge page is large (e.g. 1GB
pages).  So that locality hint is much meaningless if huge pages are used.

Signed-off-by: Peter Xu <peterx@redhat.com>