ramblock: Add ramblock_file_map()
Add a helper to do mmap() for a ramblock based on the cached informations.

A trivial thing to mention is we need to move ramblock->fd setup to be
earlier, before the ramblock_file_map() call, because it'll need to
reference the fd being mapped.  However that should not be a problem at
all, majorly because the fd won't be freed if successful, and if it failed
the fd will be freeed (or to be explicit, close()ed) by the caller.

Export it - prepare to be used outside this file.

Signed-off-by: Peter Xu <peterx@redhat.com>