migration/qapi: Replace @MigrateSetParameters with @MigrationParameters
These two structs are mostly identical besides some fields (quote from
Daniel P. Berrangé in his email):

1c1
< { 'struct': 'MigrationParameters',
---
> { 'struct': 'MigrateSetParameters',
14,16c14,16
<             '*tls-creds': 'str',
<             '*tls-hostname': 'str',
<             '*tls-authz': 'str',
---
>             '*tls-creds': 'StrOrNull',
>             '*tls-hostname': 'StrOrNull',
>             '*tls-authz': 'StrOrNull',

Here the difference is @MigrateSetParameters object would allow 'null'
values for any tls-* fields passed in, OTOH @MigrationParameters will
always be a str typed.

Here what we can do is merging the two objects into the superset, which
means we'll need to use StrOrNull for both objects.  For query parameter
command since we'll always not fetch a NULL (migration code internally
always have tls-* parameters as non-NULL, but they can be empty strings).

Here we still use the name @MigrationParameters as the final merged object
because that is a name good enough to be applied to both the QUERY or SET
operations, not the other way round - calling it @MigrateSetParameters will
be weird as an object to be returned from query-migrate-parameters QMP cmd.

This greatly deduplicates the code not only in qapi/migration.json, but
also in the generic migration code.

Signed-off-by: Peter Xu <peterx@redhat.com>