KVM: Dynamic sized kvm memslots array
Each KVMMemoryListerner maintains an array of kvm memslots.  Currently it's
statically allocated to be the max supported by the kernel.  However after
Linux commit 4fc096a99e ("KVM: Raise the maximum number of user memslots"),
the max supported memslots reported now grows to some number large enough
so that it may not be wise to always statically allocate with the max
reported.

What's worse is that QEMU kvm code still walks all the allocated memslots
entries to do any form of lookups.  It can drastically slow down all
memslot operations because each of such loop can run over 32K times on the
new kernels.

This issue was found when we observed 100% cpu consumption over the
migration thread when frequently synchronizing dirty bitmap, during which a
quick profiling shows that kvm_physical_log_clear() eats most of them
looping over empty slots.

Fix this issue by making the memslots to be allocated dynamically.  In the
case of above issue we observed, there are only ~10 memslots used.

Here the initial size was set to 16 because it should cover the basic VM
usages, so that the hope is the majority VM use case may not even need to
grow at all (e.g. if one starts qemu-system-x86_64 by default it'll consume
9 memslots), however not too large to waste memory.

Reported-by: Zhiyi Guo <zhguo@redhat.com>
Signed-off-by: Peter Xu <peterx@redhat.com>