kvmclock: Make clock_valid accurate itself
clock_valid is the variable that says "whether the QEMU ->clock is valid".
It's mostly accurate except when the current put() isn't protected against
reentrancy.  Make the put() helper check the clock validity before a real
put() applies to the kernel.

When at it, add a comment explaining the accurate meaning of clock_valid.

Signed-off-by: Peter Xu <peterx@redhat.com>