kvmclock: Move kvm sync earlier after migration
kvmclock relies on the vm state change handler to apply the correct
kvmclock data to the kernel.  This is ok but it's a bit too late for TSC
optimizations on top.  This patch moves the application of kvmclock earlier
during loading non-iterative device states.

It's achieved by simply providing a post_load() for kvmclock.

With the pre-requisite patch to allow kvmclock_put() reentrant, this should
be safe as the next vmstate notification will be void.

The goal of it is to enable kvmclock earlier than TSC, which was applied
later in cpu_synchronize_all_post_init().

Signed-off-by: Peter Xu <peterx@redhat.com>