kvmclock: Always update kvm clock with tsc
QEMU kvmclock has an optimization, where in pre_save() it will fetch again
the kvmclock just to make sure the kvmclock idle / down time is minumum.

However unlike the vmstate handler, that doesn't include a round of vcpu
TSC synchronizations.  It means when that happens, kvmclock was saved at a
different timestamp v.s. the tscs.

Logically it'll always be better to synchronize both kvmclock and tsc so
that the guest can see the two clocks change with little diff (that is what
guest OS clocksource watchdog would normally do).

So do that too in pre_save() of kvmclock, so that the tsc values are close.

To achieve this, don't check tsc valid when get: just always do it (so that
the sync in pre_save() will work even if vm stop will happen before that).

Signed-off-by: Peter Xu <peterx@redhat.com>